requirements.txt
# Core ML and Deep Learning
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Data Processing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Image Processing and Augmentation
albumentations>=1.3.0
Pillow>=8.3.0
opencv-python>=4.5.0
scikit-image>=0.19.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Experiment Tracking
tensorboard>=2.8.0
wandb>=0.13.0
mlflow>=1.28.0

# Configuration and Utilities
PyYAML>=6.0
tqdm>=4.62.0
pathlib2>=2.3.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-xdist>=2.5.0
pytest-benchmark>=4.0.0
pytest-mock>=3.7.0

# Code Quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
isort>=5.10.0
bandit>=1.7.0
pre-commit>=2.19.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.17.0
sphinx-autodoc-typehints>=1.12.0
sphinx-gallery>=0.10.0

# Performance Optimization
numba>=0.56.0

# Advanced ML Features
scikit-learn>=1.0.0
imgaug>=0.4.0
kornia>=0.6.0

# Model Deployment and Monitoring
onnx>=1.12.0
onnxruntime>=1.12.0
prometheus-client>=0.14.0

# Security
cryptography>=3.4.0

# Additional Utilities
psutil>=5.8.0
requests>=2.25.0
urllib3>=1.26.0 